refactor(mypage): 마이페이지 tab depth 평탄화
tab buttom sticky를 위해 이중탭 구조 삭제

탭 클릭 시 스크롤 상단으로 강제이동 되지 않게 임계점에 맞춰 수정

Resolves: #212
