+++ features/mypage/MyTab/TabWrapper/index.tsx
@@ -2,30 +2,37 @@
 import { useEffect, useRef } from "react";
 import PageTabs from "@/components/ui/PageTabs";
 import { useQueryParams } from "@/hooks/useQueryParams";
-import ReviewWrapper from "../../Review";
-import MeetupWrapper from "../../Meetup";
-import CreatedWrapper from "../../Created";
+import JoinedMeetingListWrapper from "../../JoinedMeetingList";
+import CreatedMeetingListWrapper from "../../CreatedMeetingList";
 import ScrollTopButton from "@/components/ui/Buttons/ScrollTopButton";
 import useScrollVisibility from "@/hooks/useScrollVisibility";
 import { cn } from "@/utils/cn";
 import useMediaQuery from "@/hooks/useMediaQuery";
+import AvailableReviewListWrapper from "../../AvailableReviewList";
+import WrittenReviewListWrapper from "../../WrittenReviewList";
 
 const MEDIA_QUERY_LG = "(min-width:1280px)";
 const MEDIA_QUERY_MD = "(min-width:744px)";
+const DEFAULT_TAB = "JoinedMeetingList";
 const THRESHOLD = {
 	lg: 20,
 	md: 320,
 	sm: 220,
 } as const;
 
-const TABS = ["meetup", "review", "created"] as const;
-type TabId = (typeof TABS)[number];
-
 const STYLE = {
 	tabWrapper: "sticky top-12 z-10 bg-gray-50 md:top-22 lg:static",
 	scroll:
-		"h-4 shadow-[0_13px_12px_rgba(0,0,0,0.08)] overflow-hidden absolute bottom-px left-0 z-0 block w-full",
+		"h-4 shadow-[0_13px_16px_rgba(0,0,0,0.08)] overflow-hidden absolute bottom-px left-0 z-0 block w-full",
 };
+
+const TABS = [
+	"JoinedMeetingList",
+	"CreatedMeetingList",
+	"AvailableReviewList",
+	"WrittenReviewList",
+] as const;
+type TabId = (typeof TABS)[number];
 
 // tab Query 검사 및 타입 가드
 function isTabId(value: string | null): value is TabId {
@@ -35,7 +42,7 @@ function isTabId(value: string | null): value is TabId {
 export default function TabWrapper() {
 	const { get, set } = useQueryParams();
 	const tabQuery = get("tab");
-	const activeTab = isTabId(tabQuery) ? tabQuery : "meetup";
+	const activeTab = isTabId(tabQuery) ? tabQuery : DEFAULT_TAB;
 	const isLg = useMediaQuery(MEDIA_QUERY_LG);
 	const isMd = useMediaQuery(MEDIA_QUERY_MD);
 	const contentRef = useRef<HTMLDivElement>(null);
@@ -54,9 +61,10 @@ export default function TabWrapper() {
 	}, [tabQuery, set]);
 
 	const tabContents: Record<TabId, React.ReactNode> = {
-		meetup: <MeetupWrapper />,
-		review: <ReviewWrapper />,
-		created: <CreatedWrapper />,
+		JoinedMeetingList: <JoinedMeetingListWrapper />,
+		CreatedMeetingList: <CreatedMeetingListWrapper />,
+		AvailableReviewList: <AvailableReviewListWrapper />,
+		WrittenReviewList: <WrittenReviewListWrapper />,
 	};
 
 	// activeTab 변경시 스크롤 초기화
@@ -64,19 +72,44 @@ export default function TabWrapper() {
 		if (isLg) {
 			contentRef.current?.scrollTo({ top: 0, behavior: "auto" });
 		} else {
-			window.scrollTo({ top: 0, behavior: "smooth" });
+			window.scrollTo({ top: isMd ? THRESHOLD.md - 20 : THRESHOLD.sm - 20, behavior: "smooth" });
 		}
-	}, [activeTab, isLg]);
+	}, [activeTab]);
 
 	return (
 		<div className="min-w-0 grow">
 			<div className={STYLE.tabWrapper}>
 				<div className={cn("relative z-1")}>
 					<div className={isVisible ? STYLE.scroll : ""} />
-					<PageTabs defaultId={activeTab} onChange={({ id }) => set({ tab: id })}>
-						<PageTabs.Item id="meetup">나의 모임</PageTabs.Item>
-						<PageTabs.Item id="review">나의 리뷰</PageTabs.Item>
-						<PageTabs.Item id="created">내가 만든 모임</PageTabs.Item>
+					<PageTabs
+						defaultId={activeTab}
+						onChange={({ id }) => {
+							set({ tab: id });
+						}}>
+						<PageTabs.Item
+							id="JoinedMeetingList"
+							className="sm:grow md:grow-0"
+							btnClassName="min-w-auto w-full">
+							참여한 모임
+						</PageTabs.Item>
+						<PageTabs.Item
+							id="CreatedMeetingList"
+							className="sm:grow md:grow-0"
+							btnClassName="min-w-auto w-full">
+							만든 모임
+						</PageTabs.Item>
+						<PageTabs.Item
+							id="AvailableReviewList"
+							className="sm:grow md:grow-0"
+							btnClassName="min-w-auto w-full">
+							리뷰 작성
+						</PageTabs.Item>
+						<PageTabs.Item
+							id="WrittenReviewList"
+							className="sm:grow md:grow-0"
+							btnClassName="min-w-auto w-full">
+							리뷰 목록
+						</PageTabs.Item>
 					</PageTabs>
 				</div>
 			</div>
